refactor: simplify edit_profile
