+++ apps/enmeshed/lib/profiles/profile/modals/modals.dart
@@ -1,2 +1,1 @@
-export 'edit_profile/edit_profile.dart';
-export 'edit_profile/edit_profile_modal.dart';
+export 'edit_profile.dart';
